// src/pages/ui/generated/catalog-api/charts.ts
// Generated by scripts/catalog-api.ts. Do not edit.

export const API = [
	{
		name: "Charts",
		props: [
			{
				name: "data",
				type: "XYPoint[]",
				required: true,
			},
			{
				name: "series",
				type: "ChartSeriesDescriptor[]",
				required: false,
			},
			{
				name: "ariaLabel",
				type: "string",
				required: false,
			},
			{
				name: "className",
				type: "string",
				required: false,
			},
		],
	},
];