// src/pages/ui/generated/catalog-api/donut.ts
// Generated by scripts/catalog-api.ts. Do not edit.

export const API = [
	{
		name: "DonutChart",
		props: [
			{
				name: "data",
				type: "NamedValue[]",
				required: true,
			},
			{
				name: "series",
				type: "ChartSeriesDescriptor[]",
				required: false,
			},
			{
				name: "ariaLabel",
				type: "string",
				required: false,
			},
			{
				name: "className",
				type: "string",
				required: false,
			},
			{
				name: "showLegend",
				type: "boolean",
				required: false,
			},
			{
				name: "valueFormatter",
				type: "(value: number) => string",
				required: false,
			},
		],
	},
];