// src/pages/ui/generated/catalog-api/gauge.ts
// Generated by scripts/catalog-api.ts. Do not edit.

export const API = [
	{
		name: "Gauge",
		props: [
			{
				name: "value",
				type: "number",
				required: true,
			},
			{
				name: "max",
				type: "number",
				required: false,
			},
			{
				name: "series",
				type: "ChartSeriesDescriptor[]",
				required: false,
			},
			{
				name: "ariaLabel",
				type: "string",
				required: false,
			},
			{
				name: "className",
				type: "string",
				required: false,
			},
			{
				name: "hideValue",
				type: "boolean",
				required: false,
			},
			{
				name: "valueFormatter",
				type: "(value: number) => string",
				required: false,
			},
		],
	},
];