// src/pages/ui/generated/catalog-api/item-list.ts
// Generated by scripts/catalog-api.ts. Do not edit.

export const API = [
	{
		name: "ItemList",
		props: [
			{
				name: "items",
				type: "ItemListItem[]",
				required: true,
			},
			{
				name: "ariaLabel",
				type: "string",
				required: false,
			},
			{
				name: "className",
				type: "string",
				required: false,
			},
		],
	},
];